Refactor to new painless package
